+++ Kernel/System/Email/SMTPTLS.pm
@@ -11,13 +11,22 @@ package Kernel::System::Email::SMTPTLS;
 use strict;
 use warnings;
 
-use Net::SSLGlue::SMTP;
+use Net::SMTP;
 
 use base qw(Kernel::System::Email::SMTP);
 
 our @ObjectDependencies = (
     'Kernel::System::Log',
 );
+
+# Use Net::SSLGlue::SMTP on systems with older Net::SMTP modules that cannot handle SMTPTLS.
+BEGIN {
+    if ( !defined &Net::SMTP::starttls ) {
+        ## nofilter(TidyAll::Plugin::OTRS::Perl::Require)
+        ## nofilter(TidyAll::Plugin::OTRS::Perl::SyntaxCheck)
+        require Net::SSLGlue::SMTP;
+    }
+}
 
 sub _Connect {
     my ( $Self, %Param ) = @_;